improvements

+++ math.go
@@ -6,7 +6,7 @@ func Add(a, b int) int {
 func Sub(a, b int) int {
 	return a - b
 }
-func Div(a, b int) int {
+func Div(a, b float64) float64 {
 	return a / b
 }
 func Mul(a, b int) int {
